+++ cpp/src/arrow/csv/converter.cc
@@ -20,6 +20,7 @@
 #include <cstring>
 #include <sstream>
 #include <string>
+#include <type_traits>
 
 #include "arrow/builder.h"
 #include "arrow/csv/parser.h"
