+++ cpp/src/arrow/csv/reader.cc
@@ -23,6 +23,8 @@
 #include <memory>
 #include <sstream>
 #include <string>
+#include <unordered_map>
+#include <utility>
 #include <vector>
 
 #include "arrow/buffer.h"
