+++ cpp/src/arrow/pretty_print-test.cc
@@ -26,12 +26,10 @@
 
 #include "arrow/array.h"
 #include "arrow/builder.h"
-#include "arrow/memory_pool.h"
 #include "arrow/pretty_print.h"
 #include "arrow/table.h"
 #include "arrow/test-util.h"
 #include "arrow/type.h"
-#include "arrow/util/decimal.h"
 
 namespace arrow {
