+++ cpp/src/arrow/pretty_print.cc
@@ -19,7 +19,6 @@
 #include <cstdint>
 #include <iostream>
 #include <memory>
-#include <sstream>
 #include <string>
 #include <type_traits>
 #include <vector>
