+++ cpp/src/arrow/pretty_print.h
@@ -21,14 +21,17 @@
 #include <ostream>
 #include <string>
 
-#include "arrow/type_fwd.h"
 #include "arrow/util/visibility.h"
 
 namespace arrow {
 
 class Array;
+class Column;
 class ChunkedArray;
+class RecordBatch;
+class Schema;
 class Status;
+class Table;
 
 struct PrettyPrintOptions {
   PrettyPrintOptions(int indent_arg, int window_arg = 10, int indent_size_arg = 2,
