+++ cpp/src/arrow/python/python-test.cc
@@ -25,6 +25,7 @@
 #include "arrow/builder.h"
 #include "arrow/table.h"
 #include "arrow/test-util.h"
+#include "arrow/util/decimal.h"
 
 #include "arrow/python/arrow_to_pandas.h"
 #include "arrow/python/decimal.h"
