+++ cpp/src/arrow/tensor.cc
@@ -17,6 +17,7 @@
 
 #include "arrow/tensor.h"
 
+#include <cstddef>
 #include <cstdint>
 #include <functional>
 #include <memory>
